requirements.txt
fastapi==0.97.0
heyoo==0.1.0
sarufi==0.1.7
python-dotenv==1.0.0
uvicorn==0.22.0
mangum==0.17.0